New: dialog to configure TEXTSEPARATOR and TEXTNUMERICLOCALE in IupMatrixEx context menu. Changed: TEXTSEPARATOR and TEXTNUMERICLOCALE attributes are now used also for all COPY* attributes in IupMatrixEx.
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@2528 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srcmatrixex/iupmatex_units.c
@@ -368,7 +368,7 @@ static int iMatrixGetUnity(const char* name, char* am_name)
   const char* s = strstr(name, "metre");
   if (s)
   {
-    int off = s - name;
+    int off = (int)(s - name);
     strcpy(am_name, name);
     strncpy(am_name+off, "meter", 5);
     return 1;
@@ -378,7 +378,7 @@ static int iMatrixGetUnity(const char* name, char* am_name)
     s = strstr(name, "litre");
     if (s)
     {
-      int off = s - name;
+      int off = (int)(s - name);
       strcpy(am_name, name);
       strncpy(am_name+off, "liter", 5);   /* don't confuse with litter */
       return 1;
